refactor(sidecar): parse blank-lines argv through the shared pass CLI DTO
